Update the version number so pip recognises the changes
When installing from a VCS source with pip it uses the version of the
package as specified in setup.py to determine if an install or upgrade
is required. Installing in editable mode is a workaround, but then it
won't execute custom build steps, which we very much need in this case!